oups

+++ mini-FTC/Robot.cpp
@@ -28,7 +28,7 @@ void Robot::run()
     wDutyCycle = 100 * ((mRobotProtocol->getEv3MotorSpeed()[2] & 0xFF)/255.0*2-1);
     //std::cout << "mMotorC: " << wDutyCycle << std::endl;
     if (mMotorC.connected()) mMotorC.set_duty_cycle_sp(wDutyCycle).run_direct();
-    
+
     wDutyCycle = 100 * ((mRobotProtocol->getEv3MotorSpeed()[3] & 0xFF)/255.0*2-1);
     //std::cout << "mMotorD: " << wDutyCycle << std::endl;
     if (mMotorD.connected()) mMotorD.set_duty_cycle_sp(wDutyCycle).run_direct();
@@ -40,7 +40,7 @@ void Robot::run()
     wEv3SensorValue[3] = mMotorD.connected() ? mMotorD.position() & 0xFF : 0;
     wEv3SensorValue[4] = mTouchSensor.connected() ? mTouchSensor.is_pressed() & 0xFF : 0;
     wEv3SensorValue[5] = mColorSensor.connected() ? mColorSensor.color() & 0xFF : 0;
-    wEv3SensorValue[6] = mUltrasonicSensor.connected() ? mUltrasonicSensor.distance_centimeters() & 0xFF : 0;
+    wEv3SensorValue[6] = mUltrasonicSensor.connected() ? (int)mUltrasonicSensor.distance_centimeters() & 0xFF : 0;
     wEv3SensorValue[7] = mGyroSensor.connected() ? mGyroSensor.angle() & 0xFF : 0;
 
     mRobotProtocol->setEv3SensorValue(wEv3SensorValue, sizeof(wEv3SensorValue));
